src/main/java/org/ninjaware/pal/rest/ControllerAdvisor.java
package org.ninjaware.pal.rest;

import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.http.converter.HttpMessageNotReadableException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.context.request.WebRequest;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

import java.util.HashMap;
import java.util.Map;

@ControllerAdvice
public class ControllerAdvisor extends ResponseEntityExceptionHandler {

    @ExceptionHandler(NotFoundException.class)
    public ResponseEntity<Object> handleNotFoundException(NotFoundException exception, WebRequest request) {
        Map<String, Object> data = new HashMap<>();
        data.put("message", exception.getMessage());
        data.put("code", "404");
        return new ResponseEntity<>(data, HttpStatus.NOT_FOUND);
    }

    @Override
    protected ResponseEntity<Object> handleHttpMessageNotReadable(HttpMessageNotReadableException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
        Map<String, Object> data = new HashMap<>();
        data.put("message", "Invalid post data supplied");
        data.put("code", "400");
        return new ResponseEntity<>(data, HttpStatus.NOT_ACCEPTABLE);
    }

}
